Move Admin files to match documents in the Starter
