+++ demo/admin/src/documents/predefinedPages/EditPredefinedPage.gql.ts
